services/src/3-backend-ai-agents/requirements.txt.txt
uagents
requests
ic-py
python-dotenv